refactor: create SearchInput component
